+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Renamed `atan_polyApprox` to `atan_polyApprox_deg9`.
   - Add template function `atan_polyApprox` which is reduced to `atan_polyApprox_deg7` or `atan_polyApprox_deg9` according to its template parameter.
   - Add template function `atan2_polyApprox` whose polynomial degree is determined at compile-time
+- Added `addSqMat` function.
 
 ## [0.4.0] - 2021/8/27